// src/app/models/ReversionTGR/ReversionTgr.interface.ts
export interface ReversionTgr {
    numeroReversion: string
    institucionId: string
    codigoBanco: string
    nombreBanco: string
    numeroTgr: string
    monto: string
    fechaPago: string
    estadoGestion: string
    motivo: string
    observacionesAprobador?: string
    fechaCreacion: string,
    transaccionBancaria: string,
    cajero: string,
    aprobadoPor: string,
    nombreAgencia: string,
    fechaHoraAprobacion: string,
    estadoTGR: string,
    nombreRazonSocial: string,
    numeroId: number
  }
  